Error handling

+++ routes/category.router.js
@@ -45,11 +45,27 @@ category.get('/:id', (req, res, next) => {
 category.post('/', (req, res, next) => {
     const { category } = req.body;
 
+    if(!category){
+        res.status(400).send({
+            message: 'Data not provided.'
+        });
+        return;
+    }
+
+    if(category?.Name.length > 128){
+        res.status(406).send({
+            message: '`Name` length is too long.'
+        });
+        return;
+    }
+
     db.post(category, (error,success) => {
         if(error){
-            res.status(500).send({
-                message: error
-            });
+            {
+                res.status(500).send({
+                    message: error.sqlMessage
+                });
+            }
         } else {
             const content = { ...category, Id: success.insertId };
             res.status(201).json({
@@ -67,6 +83,13 @@ category.post('/', (req, res, next) => {
 category.delete('/:id', (req, res, next) => {
     const { id } = req.params;
 
+    if(!id){
+        res.status(400).send({
+            message: 'ID not provided.'
+        });
+        return;
+    }
+    
     db.delete(id, (error, success) => {
         if(error){
             res.status(500).send({
